+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Add new properties ``pybed.BedFrame.has_chr_prefix``, ``pycov.CovFrame.has_chr_prefix``, ``pyvcf.VcfFrame.has_chr_prefix``.
 * Add new method :meth:`pyvcf.slice`.
 * Add new method :meth:`pyvcf.VcfFrame.from_string`.
+* Remove ``nrows`` argument from :meth:`pyvcf.VcfFrame.from_file` method.
 
 0.25.0 (2021-10-09)
 -------------------